Refactor AnimalKingdomASTCreationTests
